% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_functions.R
\name{ComputeErrorRate}
\alias{ComputeErrorRate}
\title{Computes error rate given estimated and real cluster ids.}
\usage{
ComputeErrorRate(zhat, z)
}
\arguments{
\item{zhat}{vector. Estimated cluster ids.}

\item{z}{vector. True cluster ids.}
}
\description{
Computes error rate given estimated and real cluster ids.
}
