% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_functions.R
\name{GenerateSyntheticData}
\alias{GenerateSyntheticData}
\title{Generates synthetic data.}
\usage{
GenerateSyntheticData(n, p, s, k, signal_strength, noise_type)
}
\arguments{
\item{n}{int. Number of observations.}

\item{p}{int. Number of features.}

\item{s}{int. Number of informative features.}

\item{k}{int. Number of clusters.}

\item{signal_strength}{float. Signal strength.}

\item{noise_type}{character. Noise type. Must be either "gaussian" or "t2".}
}
\value{
list. The result contains two attributes: $data is the data matrix, $labels contain the cluster ids.
}
\description{
This function generates synthetic clustering data.

It generates synthetic by the following steps:

  1. generates cluster centers with s informative and p total features according to the following steps:
     1). generates orthonormal matrix of dimension k by s.
     2). multiplies the matrix by signal_strength.
     3). binds (p-s) zero columns to the above matrix.

  2. randomly generates n by k one-hot cluster assignments.

  3. generates n by p signal matrix and add scaled standard gaussian or t2 noise.
}
\examples{
GenerateSyntheticData(n=10, p=10, s=5, k=2, signal_strength=1, noise_type="gaussian")
}
\references{
T. Liu, Y. Lu, B. Zhu, H. Zhao (2021). High-dimensional Clustering via Feature Selection with
Applications to Single Cell RNA-seq Data.
}
