% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_functions.R
\name{LloydIteration}
\alias{LloydIteration}
\title{Conducts Lloyd iterations}
\usage{
LloydIteration(data, num_clusters, init_cluster_ids, num_iterations)
}
\arguments{
\item{data}{matrix. Input data matrix for clustering.}

\item{num_clusters}{int. Number of clusters.}

\item{init_cluster_ids}{vector. The initial clusters assignment.}

\item{num_iterations}{int. Number of Lloyd iterations.}
}
\value{
A vector containing the cluster assignment after iterations.
}
\description{
This function conducts Lloyd iterations from an initial clusters assignment.
}
