% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_functions.R
\name{SpectralClusterFeatureSelection}
\alias{SpectralClusterFeatureSelection}
\title{Conducts spectral clustering followed by feature selection and another clustering.}
\usage{
SpectralClusterFeatureSelection(data, num_clusters, init_cluster_ids, use_lloyd_iteration)
}
\arguments{
\item{data}{matrix. Input data matrix for clustering.}

\item{num_clusters}{int. Number of clusters.}

\item{init_cluster_ids}{vector. The initial clusters assignment. If NULL, the initial guess will 
be obtained by spectral clustering. Otherwise, step 1 is skipped and the feature selection will be
applied based on the initial guess.}

\item{use_lloyd_iteration}{bool. If TRUE, conduct Lloyd iteration at step 3. Otherwise, no Lloyd iteration.}
}
\value{
A vector containing the estimated cluster assignment.
}
\description{
This function implements the main algorithm of SCFS.

It conducts clustering by the following steps:

  1. conducts spectral clustering to obtain initial guess of cluster assignments.
  
  2. uses the initial guess to compute R^2 and select informative features.
  
  3. conducts spectral clustering again (SCFS1) followed by Lloyd iteration (optionally for SCFS2).
}
\examples{
synthetic_data <- GenerateSyntheticData(n=10, p=10, s=5, k=2, signal_strength=1, noise_type="gaussian")
scfs <- SpectralClusterFeatureSelection(data=synthetic_data$data, num_clusters=2, init_cluster_ids=NULL, use_lloyd_iteration=FALSE)
}
\references{
T. Liu, Y. Lu, B. Zhu, H. Zhao (2021). High-dimensional Clustering via Feature Selection with
Applications to Single Cell RNA-seq Data.
}
