% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_functions.R
\name{SpectralClustering}
\alias{SpectralClustering}
\title{Conducts spectral clustering algorithm.}
\usage{
SpectralClustering(data, num_clusters)
}
\arguments{
\item{data}{matrix. Input data matrix for clustering.}

\item{num_clusters}{int. Number of clusters.}
}
\value{
A vector containing the cluster assignment after iterations.
}
\description{
This function implements spectral clustering algorithm.

It conducts k-means on top k left singular vectors of the data matrix.
}
\examples{
synthetic_data <- GenerateSyntheticData(n=10, p=10, s=5, k=2, signal_strength=1, noise_type="gaussian")
label.est <- SpectralClustering(synthetic_data$data, 2)
}
\references{
T. Liu, Y. Lu, B. Zhu, H. Zhao (2021). High-dimensional Clustering via Feature Selection with
Applications to Single Cell RNA-seq Data.
}
